+++ src/js/popup/pages/account/buy.jsx
