+++ src/js/popup/pages/account/details.jsx
