+++ src/js/popup/pages/account/index.jsx
