+++ src/js/popup/pages/wallet/chooseNetwork.jsx
