+++ src/js/popup/pages/wallet/index.jsx
